Move Servlet Mocks to Web
Issue gh-13551
