+++ web/src/test/java/org/springframework/security/web/servlet/TestMockHttpServletMappings.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.security.config;
+package org.springframework.security.web.servlet;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.MappingMatch;
